Cleanup abstractions.
