+++ crates/provider/src/traits.rs
@@ -136,29 +136,12 @@ where
     // Low-level methods
     // ========================================================================
 
-    /// Makes a call request while handling seismic specific aspects
-    /// e.g. sending signed call requests
+    /// Low-level seismic call. Sends an `eth_call` RPC request.
+    ///
+    /// The default implementation sends the request directly without filling
+    /// or decryption. [`ResponseDecryptProvider`](crate::decrypt::ResponseDecryptProvider)
+    /// overrides this to fill, send, and decrypt.
     async fn seismic_call(&self, tx: SendableTx<N>) -> TransportResult<Bytes> {
-        self.call_conditionally_signed(tx).await
-    }
-
-    /// Whether the input data should be encrypted.
-    /// If it's not a seismic tx, don't encrypt.
-    /// If it is, only encrypt if it's non-empty.
-    fn should_encrypt_input<B: TransactionBuilder<N>>(&self, tx: &B) -> bool {
-        if !N::is_seismic_tx_type(tx.output_tx_type()) {
-            return false;
-        }
-        tx.input().map_or(false, |input| !input.is_empty())
-    }
-
-    /// Get the PublicKey of the enclave
-    async fn get_tee_pubkey(&self) -> TransportResult<PublicKey> {
-        seismic_alloy_network::fetch_tee_pubkey(self).await
-    }
-
-    /// Makes a call request, perhaps making the call signed depending on the input type
-    async fn call_conditionally_signed(&self, tx: SendableTx<N>) -> TransportResult<Bytes> {
         match tx {
             SendableTx::Builder(builder) => {
                 let output: Bytes = self.client().request("eth_call", (builder.clone(),)).await?;
@@ -171,15 +154,22 @@ where
             }
         }
     }
+
+    /// Get the PublicKey of the enclave.
+    async fn get_tee_pubkey(&self) -> TransportResult<PublicKey> {
+        seismic_alloy_network::fetch_tee_pubkey(self).await
+    }
 }
+
+// ---------------------------------------------------------------------------
+// Blanket impls
+// ---------------------------------------------------------------------------
 
 impl SeismicProviderExt<SeismicReth> for RootProvider<SeismicReth> {}
 impl SeismicProviderExt<SeismicFoundry> for RootProvider<SeismicFoundry> {}
 
 /// Blanket impl so `&T: SeismicProviderExt` when `T: SeismicProviderExt`.
-/// This mirrors alloy's blanket `Provider` impl for `&T` and is needed so that
-/// `#[sol(rpc)]`-generated contract types (which store `&provider`) work with
-/// `.seismic()`.
+/// Needed for `#[sol(rpc)]`-generated contract types which store `&provider`.
 #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
 impl<T, N> SeismicProviderExt<N> for &T
@@ -194,6 +184,11 @@ where
     }
 }
 
+/// Blanket impl for `FillProvider` — fills the transaction then delegates to `RootProvider`.
+///
+/// This is the path used by unsigned providers (which are bare `FillProvider`s).
+/// Signed providers use [`ResponseDecryptProvider`](crate::decrypt::ResponseDecryptProvider)
+/// instead, which fills, sends, AND decrypts.
 #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
 impl<F, P, N> SeismicProviderExt<N> for FillProvider<F, P, N>
@@ -206,15 +201,16 @@ where
     RootProvider<N>: SeismicProviderExt<N>,
 {
     async fn seismic_call(&self, tx: SendableTx<N>) -> TransportResult<Bytes> {
-        // Fill the transaction
-        let builder = tx.as_builder().unwrap().clone();
+        let builder = match tx {
+            SendableTx::Builder(b) => b,
+            SendableTx::Envelope(_) => {
+                return Err(TransportErrorKind::custom_str(
+                    "FillProvider::seismic_call received an Envelope; expected a Builder",
+                ));
+            }
+        };
 
-        let built_tx = self.fill(builder).await?;
-
-        // self.inner is not public for FillProvider.
-        // However, for our use cases, self.inner is the RootProvider,
-        // so we get it this hacky way
-        let inner = self.root();
-        SeismicProviderExt::seismic_call(inner, built_tx).await
+        let filled_tx = self.fill(builder).await?;
+        SeismicProviderExt::seismic_call(self.root(), filled_tx).await
     }
 }
